% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potenciaRecursiva.R
\name{potenciaRecursiva}
\alias{potenciaRecursiva}
\title{Function that returns the power of a number}
\usage{
potenciaRecursiva(x, expo = 1, logScale = FALSE)
}
\arguments{
\item{x}{Base}

\item{expo}{expoente em ingles}

\item{logScale}{Whether the result should be returned in the log scale. If TRUE
\code{x} must be provided in the log scale as well.}
}
\value{
The result of the calculations.
}
\description{
Return the power in the natural or log scale. Function recursion exercise
}
\examples{
potenciaRecursiva(2, 6)
}
